requirements.txt
# Core Juniper PyEZ dependencies
junos-eznc>=2.7.0
ncclient>=0.6.15

# Configuration and templating
PyYAML>=6.0
Jinja2>=3.1.0

# Data analysis and comparison
deepdiff>=6.7.0
tabulate>=0.9.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# WebSocket support
websockets>=12.0

# Terminal User Interface
textual>=0.50.0
rich>=13.7.0
aiohttp>=3.9.0

# Network topology and analysis
networkx>=3.2.0

# XML processing
lxml>=5.0.0

# System monitoring
psutil>=5.9.0

# HTTP client for testing
requests>=2.31.0

# Optional: JSNAPy for validation (may require additional system dependencies)
 jsnapy

# Optional: Advanced reporting
# reportlab>=4.1.0
# matplotlib>=3.8.0

# Development and testing (compatible versions)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality tools
black>=24.0.0
flake8>=7.0.0

# Note: mypy, bandit, and safety are optional development tools
# Uncomment if needed for development:
# mypy>=1.8.0
# bandit>=1.7.0
# safety>=3.0.0
